add function to parse string for links, create parsing helpers file
