+++ src/helpers/index.js
@@ -1,5 +1,6 @@
 export * from './prop_checker';
 export * from './sidebar';
+export * from './parsing_helpers';
 
 // create comma seperated number
 export const formatNumber = (number = 301, shouldAbbreviate = false) => {
@@ -51,54 +52,6 @@ export const createUrlParams = obj => (
   	}
   }).filter( str => str !== undefined ).join('&')
 );
-
-// parse video duration from api call
-export const parseDuration = str => {
-  let values = str.split(/[A-Z]+/);
-  values = values.filter( val => val !== '');
-  values = values.map( (val, idx) => {
-  	if(idx === 0) {
-  		return val;
-  	} else {
-  		if(val.length === 2) {
-  			return val;
-  		} else {
-  			return '0' + val;
-  		}
-  	}
-  });
-
-  if (values.length === 1) {
-    values.unshift('0');
-  }
-  return values.join(':');
-};
-
-// return written version of date
-export const parseDate = date => {
-  const months = {
-    0: "Jan",
-    1: "Feb",
-    2: "Mar",
-    3: "Apr",
-    4: "May",
-    5: "June",
-    6: "July",
-    7: "Aug",
-    8: "Sept",
-    9: "Oct",
-    10: "Nov",
-    11: "Dec"
-  };
-  date = new Date(date);
-  let newDate = "";
-
-  newDate += months[date.getMonth()] + " ";
-  newDate += date.getDate() + ", ";
-  newDate += date.getFullYear();
-
-  return newDate;
-};
 
 export const timeFromNow = date => {
   let durationsInSeconds = {
